Also accept byte slice in scan on generated code
I've seen that Scan gets passed a slice of bytes coming from the
database layer. This doesn't match the string in the type switch
statement, but makes sense to pass into the json unmarshal
functionality directly.

This commit allows a []byte to be passed into the Scan method and is
handled very similarly to the string, it doesn't need to be cast though.