Define a jsonenum-generated error for invalid values
This will replace the generic fmt.Errorf error currently used. The new
'InvalidValueError' method provides a way for clients to detect this
type of error through an interface-satisfaction check. Including the bad
value is useful to maintain parity with what is currently interpolated
into the fmt.Errorf errors that are being used instead.